### 2018-05-23 #751 @whedon
[PRE REVIEW]: Neutron Imaging Normalization Library
**Submitting author:** @jeanbilheux (<a href="http://orcid.org/0000-0003-2172-6487">Jean Bilheux</a>)
**Repository:** <a href="https://github.com/scikit-beam/NeuNorm" target ="_blank">https://github.com/scikit-beam/NeuNorm</a>
**Version:** 1.4.0
