### 2018-05-23 57eea7a
Update meta.yaml
Fixed LICENSE.txt file link
