### 2018-05-23 210a488
added license.txt to list of files to keep
