+++ conda-recipes/meta.yaml
@@ -1,5 +1,5 @@
-{% set version = "1.4.0" %}
-{% set git_rev = "1.4.0" %}
+{% set version = "1.4.1" %}
+{% set git_rev = "1.4.1" %}
 
 package:
   name: neunorm
