### 2018-05-23 1caf2d2
update list of files to ignore and updated version for pypi
