+++ README.md
@@ -1,7 +1,7 @@
 [![pypi version](https://img.shields.io/pypi/v/neunorm.svg)](https://pypi.python.org/pypi/neunorm)
 [![Anaconda-Server Badge](https://anaconda.org/neutronimaging/neunorm/badges/version.svg)](https://anaconda.org/neutronimaging/neunorm)
 [![Build Status](https://travis-ci.org/scikit-beam/NeuNorm.svg?branch=master)](https://travis-ci.org/scikit-beam/NeuNorm)
-[![codecov](https://codecov.io/gh/ornlneutronimaging/NeuNorm/branch/master/graph/badge.svg)](https://codecov.io/gh/ornlneutronimaging/NeuNorm)
+[![codecov](https://codecov.io/gh/scikit-beam/NeuNorm/branch/master/graph/badge.svg)](https://codecov.io/gh/scikit-beam/NeuNorm)
 [![Documentation Status](https://readthedocs.org/projects/neunorm/badge/?version=latest)](http://neunorm.readthedocs.io/en/latest/?badge=latest)
 [![Anaconda-Server Badge](https://anaconda.org/neutronimaging/neunorm/badges/latest_release_date.svg)](https://anaconda.org/neutronimaging/neunorm)
 [![Anaconda-Server Badge](https://anaconda.org/neutronimaging/neunorm/badges/downloads.svg)](https://anaconda.org/neutronimaging/neunorm)
