### 2018-06-25 #751 @arfon
@JeanBilheux - could you take a look at this list of reviewers and suggest a few that might be good candidates to review this package? https://bit.ly/joss-reviewers
