### 2018-07-28 #815 @stuartcampbell
@arfon Sorry for the delay in getting back to this.

I was just reading the conflict of interest, and I notice that it has a definition of recent as 4 years… I used to be in the same group as Jean 2 years ago, but I was never involved in any of the work he was doing in the area related to this publication.  This will be the first time I will have actually seen or looked at any of this code/work.

Also, I am an owner of the ‘scikit-beam’ repo that this software is being hosted under, but this is just an umbrella org to try and co-locate any python based software from synchrotron and neutron sources.

I just wanted to mention these in case it rules me out of being a reviewer.
