### 2018-07-31 scikit-beam/NeuNorm#4 @stuartcampbell
Github Releases and Tags are not up to date
Hi Jean,
I am performing a review of NeuNorm for JOSS under the ticket (https://github.com/openjournals/joss-reviews/issues/815)

I notice that the version submitted to JOSS is 1.4.0, but I can't seem to find either a Github or tag for this release (or the newer 1.4.1) that is on PyPI.

I'll create separate issues here for any other comments and suggestions if that's ok ?

Cheers
Stu
