### 2018-07-31 scikit-beam/NeuNorm#4 @JeanBilheux
Sorry about that. I just created the 1.4.1 tag. Bad habits I have of creating pypi and conda versions without creating a tag first. I will try not to do it again.
