### 2018-08-02 #815 @stuartcampbell
@JeanBilheux - I've think I'm done with the review.  All in all looks good, I would just say to just add some simple contributing guidelines and update the references in the paper to include the software you are making use of (basically what you already have in you .bib file) - Then I think we are good to go!  Cheers
Stu
