### 2018-08-03 141cd4a
Updated list of references
