### 2018-08-03 0328f18
Updated README.md by adding Meta, references and acknowledgements
