### 2018-08-03 0229b19
Create CONTRIBUTING.md
