+++ CONTRIBUTING.md
@@ -1,0 +1,18 @@
+# Contributing to NeuNorm
+
+## Reporting issues
+
+When reporting issues please include as much detail as possible about your
+operating system, NeuNorm version and python version. Whenever possible, please
+also include a brief, self-contained code example that demonstrates the problem.
+
+## Contributing code
+
+Thanks for your interest in contributing code to NeuNorm!
+
+We are currently working on new features such as:
+ - use of a mask to block pixels and not allow normalization of those pixels
+ - speed up loading of data by working in parallel
+ 
+For any contribution you would like to add, please fork the NeuNorm project and use the pull request to bring them back to master.
+
