### 2018-08-03 #815 @JeanBilheux
I updated the references in the README.md as well as the paper.md files. I added a CONTRIBUTING.md file as well.
