### 2018-08-03 #815 @JeanBilheux
With all those changes, should I make a new release version 1.4.2 ?
