### 2018-08-03 #815 @arfon
> With all those changes, should I make a new release version 1.4.2 ?

Yes, that sounds like a good idea.
