### 2018-08-06 #815 @stuartcampbell
@arfon I'm happy with the changes made and have finished my review
