### 2018-08-06 #815 @arfon
Thanks @stuartcampbell! @JeanBilheux - At this point could you make an archive of the reviewed software in Zenodo/figshare/other service and update this thread with the DOI of the archive? I can then move forward with accepting the submission.
